fix deleted

+++ src/main/java/com/mkyong/common/action/FileUploadAction.java
@@ -44,12 +44,12 @@ public class FileUploadAction extends Action{
 	        System.out.println("Server path: " +filePath);
 	        File newFile = new File(filePath, fileName);
               
-	        /*if(!newFile.exists()){
+	        if(!newFile.exists()){
 	          FileOutputStream fos = new FileOutputStream(newFile);
 	          fos.write(file.getFileData());
 	          fos.flush();
-	          fos.close()
-	        }  */
+	          fos.close();
+	        }  
 	        
 	        request.setAttribute("uploadedFilePath",newFile.getAbsoluteFile());
 	        request.setAttribute("uploadedFileName",newFile.getName());
@@ -61,7 +61,7 @@ public class FileUploadAction extends Action{
             String readfile = fileUploadForm.getReadfile();
             if(!("").equals(readfile)){
                 File f=new File(filePath, readfile);
-                /*FileInputStream fs = new FileInputStream(f);
+                FileInputStream fs = new FileInputStream(f);
                 InputStreamReader in = new InputStreamReader(fs);
                 BufferedReader br = new BufferedReader(in);
         	
@@ -69,7 +69,7 @@ public class FileUploadAction extends Action{
                 while ((line = br.readLine()) != null) {
                     output = output + line + "\n<br>";
                 } 
-                br.close();*/
+                br.close();
             }
            
             request.setAttribute("readfile",output);
